Refactor Engine class and enhance resource management
- Removed shader compilation methods from Engine class and integrated ShaderCompiler for better separation of concerns.
- Introduced InputManager for handling mouse and keyboard input, improving input handling structure.
- Updated Engine constructor to utilize new ResourceManager for managing GPU resources and loading models/textures.
- Enhanced scene loading functionality with SceneLoader for better organization of entity creation from scene files.
- Cleaned up unused code and improved overall readability and maintainability of the engine codebase.
