+++ engine/engine.cpp
@@ -1,53 +1,42 @@
-#define STB_IMAGE_IMPLEMENTATION
-#define TINYOBJLOADER_IMPLEMENTATION
 #include "engine.hpp"
 
 #include <array>
 #include <cstdlib>
-#include <tiny_obj_loader.h>
 #include <fstream>
 #include <sstream>
 
+// Include all the subsystem headers
+#include "shaders/shader_compiler.hpp"
+#include "input/input_manager.hpp"
+#include "resources/resource_manager.hpp"
+#include "scene/scene_loader.hpp"
+#include "rendering/vulkan_init.hpp"
+#include "rendering/vulkan_resources.hpp"
+#include "rendering/vulkan_renderer.hpp"
+
 namespace impgine {
 
-
-VkResult CreateDebugUtilsMessengerEXT(VkInstance instance,
-                                      const VkDebugUtilsMessengerCreateInfoEXT* pCreateInfo,
-                                      const VkAllocationCallbacks* pAllocator,
-                                      VkDebugUtilsMessengerEXT* pDebugMessenger) {
-    auto func = (PFN_vkCreateDebugUtilsMessengerEXT)vkGetInstanceProcAddr(
-        instance, "vkCreateDebugUtilsMessengerEXT");
-    if (func != nullptr) {
-        return func(instance, pCreateInfo, pAllocator, pDebugMessenger);
-    } else {
-        return VK_ERROR_EXTENSION_NOT_PRESENT;
-    }
-}
-
-void DestroyDebugUtilsMessengerEXT(VkInstance instance, VkDebugUtilsMessengerEXT debugMessenger,
-                                   const VkAllocationCallbacks* pAllocator) {
-    auto func = (PFN_vkDestroyDebugUtilsMessengerEXT)vkGetInstanceProcAddr(
-        instance, "vkDestroyDebugUtilsMessengerEXT");
-    if (func != nullptr) {
-        func(instance, debugMessenger, pAllocator);
-    }
-}
+// External function to set ResourceManager globals
+extern void setResourceManagerGlobals(VkCommandPool commandPool, VkQueue graphicsQueue);
 
 Engine::Engine() {
     // Load project configuration
     project = Project::loadProject("example/project.imp");
 
     // Compile shaders before initializing Vulkan
-    compileShaders();
+    ShaderCompiler::compileAllShaders(project);
 
     window = std::make_unique<Window>(project.windowWidth, project.windowHeight, project.windowTitle);
     window->setUserPointer(this);
     window->setFramebufferSizeCallback(framebufferResizeCallback);
     glfwSetMouseButtonCallback(window->getGLFWWindow(), mouseButtonCallback);
+
     initVulkan();
 }
 
-Engine::~Engine() { cleanup(); }
+Engine::~Engine() {
+    cleanup();
+}
 
 void Engine::run() {
     std::cout << "Welcome to Impgine!\n";
@@ -61,18 +50,19 @@ void Engine::run() {
         for (const auto& component : ECSRegistry::getRegistry().getComponents(entity)) {
             std::cout << "Component: " << component.first.name() << std::endl;
         }
-
     }
 
     mainLoop();
 }
 
 void Engine::initVulkan() {
-    createInstance();
-    setupDebugMessenger();
-    createSurface();
-    pickPhysicalDevice();
-    createLogicalDevice();
+    // Use Vulkan init functions
+    createInstance(instance, validationLayers, enableValidationLayers);
+    setupDebugMessenger(instance, debugMessenger, enableValidationLayers);
+    createSurface(instance, surface, window.get());
+    pickPhysicalDevice(instance, physicalDevice, surface, deviceExtensions, msaaSamples, needsPortabilitySubset);
+    createLogicalDevice(physicalDevice, device, graphicsQueue, presentQueue, surface,
+                       deviceExtensions, validationLayers, enableValidationLayers, needsPortabilitySubset);
 
     createECSRegistry();
     
@@ -82,61 +72,69 @@ void Engine::initVulkan() {
 
     swapChain = std::make_unique<SwapChain>(device, physicalDevice, surface, *window);
 
-    createCommandPool();
+    // Create command pool using vulkan_resources
+    createCommandPool(device, physicalDevice, surface, commandPool);
+
+    // Create subsystems
+    inputManager = new InputManager(window.get(), &camera);
+    resourceManager = new ResourceManager(device, physicalDevice);
+
+    // Configure fallback texture path
+    resourceManager->setFallbackTexturePath(project.getFullPath("engine/textures/texture.jpg"));
+
+    // Set global resource manager context
+    setResourceManagerGlobals(commandPool, graphicsQueue);
+
     // Defer content loading to loadScene
-    createColorResources();
-    createDepthResources();
-    createRenderPass();
-    createFramebuffers();
+    createColorResources(device, physicalDevice, swapChain->getSwapChainImageFormat(),
+                        swapChain->getSwapChainExtent(), msaaSamples,
+                        colorImage, colorImageMemory, colorImageView);
+    createDepthResources(device, physicalDevice, swapChain->getSwapChainExtent(),
+                        msaaSamples, depthImage, depthImageMemory, depthImageView);
+    createRenderPass(device, physicalDevice, swapChain->getSwapChainImageFormat(),
+                    msaaSamples, renderPass);
+    std::vector<VkImageView> imageViews;
+    for (size_t i = 0; i < swapChain->imageCount(); i++) {
+        imageViews.push_back(swapChain->getImageView(i));
+    }
+    createFramebuffers(device, renderPass, swapChain->getSwapChainExtent(),
+                      imageViews, colorImageView, depthImageView,
+                      swapChainFramebuffers);
 
     // Pipeline will be created after scene resources are ready
 }
 
 void Engine::createECSRegistry() {
     auto& reg = ECSRegistry::getRegistry();
-    (void)reg; 
+    (void)reg;
 }
 
-    void Engine::mainLoop() {
-        auto lastTime = std::chrono::high_resolution_clock::now();
-        
+void Engine::mainLoop() {
+    auto lastTime = std::chrono::high_resolution_clock::now();
+    
     while (!window->shouldClose()) {
         auto currentTime = std::chrono::high_resolution_clock::now();
         float deltaTime = std::chrono::duration<float>(currentTime - lastTime).count();
         lastTime = currentTime;
         
         window->pollEvents();
-        handleUIInput();
-        processInput(deltaTime);
-        handleMouseMovement();
+        inputManager->handleUIInput(uiRenderer.get(), uiWindows);
+        inputManager->processInput(deltaTime);
+        inputManager->handleMouseMovement();
         drawFrame();
     }
     vkDeviceWaitIdle(device);
 }
 
 void Engine::cleanupSwapChain() {
-    vkDestroyImageView(device, colorImageView, nullptr);
-    vkDestroyImage(device, colorImage, nullptr);
-    vkFreeMemory(device, colorImageMemory, nullptr);
+    impgine::cleanupSwapChain(device, colorImageView, colorImage, colorImageMemory,
+                              depthImageView, depthImage, depthImageMemory,
+                              swapChainFramebuffers, pipeline.get(), renderPass, swapChain.get());
 
-    vkDestroyImageView(device, depthImageView, nullptr);
-    vkDestroyImage(device, depthImage, nullptr);
-    vkFreeMemory(device, depthImageMemory, nullptr);
-
-    for (auto framebuffer : swapChainFramebuffers) {
-        vkDestroyFramebuffer(device, framebuffer, nullptr);
-    }
-
-    if (pipeline) {
-        pipeline.reset();
-    }
-
-    vkDestroyRenderPass(device, renderPass, nullptr);
-
-    if (swapChain) {
-        swapChain.reset();
-    }
-    }
+    // Reset unique_ptrs after cleanup (they will call destructors)
+    pipeline.reset();
+    swapChain.reset();
+}
 
 void Engine::initializeUnityLayout() {
     uiWindows.clear();
@@ -200,7 +198,7 @@ void Engine::initializeUnityLayout() {
         }
     }
     std::cout << "============================\n" << std::endl;
-}
+    }
 
 void Engine::cleanup() {
     // Destroy UI resources before destroying device
@@ -209,232 +207,76 @@ void Engine::cleanup() {
     }
     cleanupSwapChain();
 
-    // Cleanup all mesh resources
-    for (auto& [path, meshResources] : meshCache) {
-        vkDestroySampler(device, meshResources.textureSampler, nullptr);
-        vkDestroyImageView(device, meshResources.textureImageView, nullptr);
-        vkDestroyImage(device, meshResources.textureImage, nullptr);
-        vkFreeMemory(device, meshResources.textureImageMemory, nullptr);
-        vkDestroyBuffer(device, meshResources.indexBuffer, nullptr);
-        vkFreeMemory(device, meshResources.indexBufferMemory, nullptr);
-        vkDestroyBuffer(device, meshResources.vertexBuffer, nullptr);
-        vkFreeMemory(device, meshResources.vertexBufferMemory, nullptr);
+    // Cleanup all mesh resources via ResourceManager
+    if (resourceManager) {
+        for (auto& [path, meshResources] : resourceManager->getMeshCache()) {
+            vkDestroySampler(device, meshResources.textureSampler, nullptr);
+            vkDestroyImageView(device, meshResources.textureImageView, nullptr);
+            vkDestroyImage(device, meshResources.textureImage, nullptr);
+            vkFreeMemory(device, meshResources.textureImageMemory, nullptr);
+            vkDestroyBuffer(device, meshResources.indexBuffer, nullptr);
+            vkFreeMemory(device, meshResources.indexBufferMemory, nullptr);
+            vkDestroyBuffer(device, meshResources.vertexBuffer, nullptr);
+            vkFreeMemory(device, meshResources.vertexBufferMemory, nullptr);
+        }
     }
-    meshCache.clear();
+
+    vkDestroyDescriptorPool(device, descriptorPool, nullptr);
 
     for (size_t i = 0; i < uniformBuffers.size(); i++) {
         vkDestroyBuffer(device, uniformBuffers[i], nullptr);
         vkFreeMemory(device, uniformBuffersMemory[i], nullptr);
     }
 
-    vkDestroyDescriptorPool(device, descriptorPool, nullptr);
     vkDestroyDescriptorSetLayout(device, descriptorSetLayout, nullptr);
 
-    if (pipelineLayout != VK_NULL_HANDLE) {
         vkDestroyPipelineLayout(device, pipelineLayout, nullptr);
-    }
 
-    if (commandPool != VK_NULL_HANDLE) {
         vkDestroyCommandPool(device, commandPool, nullptr);
-    }
 
-    if (device != VK_NULL_HANDLE) {
         vkDestroyDevice(device, nullptr);
-    }
 
     if (enableValidationLayers) {
         DestroyDebugUtilsMessengerEXT(instance, debugMessenger, nullptr);
     }
 
-    if (surface != VK_NULL_HANDLE) {
         vkDestroySurfaceKHR(instance, surface, nullptr);
-    }
-
-    if (instance != VK_NULL_HANDLE) {
         vkDestroyInstance(instance, nullptr);
-    }
-}
 
-void Engine::createInstance() {
-    if (enableValidationLayers && !checkValidationLayerSupport()) {
-        throw std::runtime_error("validation layers requested, but not available!");
-    }
-
-    VkApplicationInfo appInfo{};
-    appInfo.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO;
-    appInfo.pApplicationName = "Impgine Application";
-    appInfo.applicationVersion = VK_MAKE_VERSION(1, 0, 0);
-    appInfo.pEngineName = "Impgine";
-    appInfo.engineVersion = VK_MAKE_VERSION(1, 0, 0);
-    appInfo.apiVersion = VK_API_VERSION_1_0;
-
-    VkInstanceCreateInfo createInfo{};
-    createInfo.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO;
-    createInfo.pApplicationInfo = &appInfo;
-
-    auto extensions = getRequiredExtensions();
-    extensions.push_back(VK_KHR_PORTABILITY_ENUMERATION_EXTENSION_NAME);
-
-    createInfo.enabledExtensionCount = static_cast<uint32_t>(extensions.size());
-    createInfo.ppEnabledExtensionNames = extensions.data();
-    createInfo.flags |= VK_INSTANCE_CREATE_ENUMERATE_PORTABILITY_BIT_KHR;
-
-    VkDebugUtilsMessengerCreateInfoEXT debugCreateInfo{};
-    if (enableValidationLayers) {
-        createInfo.enabledLayerCount = static_cast<uint32_t>(validationLayers.size());
-        createInfo.ppEnabledLayerNames = validationLayers.data();
-
-        populateDebugMessengerCreateInfo(debugCreateInfo);
-        createInfo.pNext = (VkDebugUtilsMessengerCreateInfoEXT*)&debugCreateInfo;
-    } else {
-        createInfo.enabledLayerCount = 0;
-        createInfo.pNext = nullptr;
-    }
-
-    if (vkCreateInstance(&createInfo, nullptr, &instance) != VK_SUCCESS) {
-        throw std::runtime_error("failed to create instance!");
-    }
+    // Clean up subsystems
+    delete inputManager;
+    delete resourceManager;
 }
 
 void Engine::loadScene(const std::string& scenePath) {
-    // Minimal YAML-like scene format inspired by Unity text:
-    // --- !imp!Entity &id
-    // name: MyEntity
-    // Transform:
-    //   position: {x: 0, y: 0, z: 0}
-    //   rotation: {x: 0, y: 0, z: 0}
-    //   scale: {x: 1, y: 1, z: 1}
-    // Model:
-    //   modelPath: models/viking_room.obj
-    //   texturePath: textures/viking_room.png
-    //   color: {r: 1, g: 1, b: 1}
+    // Use SceneLoader to load scene data into ECS
+    SceneLoader::loadScene(scenePath, project);
 
-    std::ifstream file(scenePath);
-    if (!file.is_open()) {
-        std::cerr << "Could not open scene file: " << scenePath << ". Using defaults.\n";
-    }
-
-    struct PendingEntity {
-        std::optional<std::string> tag;
-        glm::vec3 position {0.0f};
-        glm::vec3 rotation {0.0f};
-        glm::vec3 scale {1.0f};
-        std::string modelPath;
-        std::string texturePath;
-        glm::vec3 color {1.0f, 1.0f, 1.0f};
-    };
-
-    std::vector<PendingEntity> entitiesToCreate;
-    PendingEntity current; 
-    bool inEntityBlock = false;
-    enum class Section { None, Transform, MeshRenderer3D };
-    Section section = Section::None;
-
-    auto trim = [](std::string s) {
-        size_t a = s.find_first_not_of(" \t\r\n");
-        size_t b = s.find_last_not_of(" \t\r\n");
-        if (a == std::string::npos) return std::string();
-        return s.substr(a, b - a + 1);
-    };
-
-    if (file.is_open()) {
-        std::string line;
-        while (std::getline(file, line)) {
-            line = trim(line);
-            if (line.empty() || line[0] == '#') continue;
-            if (line.rfind("%", 0) == 0) continue; // header like %IMP 1.0
-
-            if (line.rfind("---", 0) == 0) {
-                // Start of a new entity block if tagged as !imp!Entity
-                if (line.find("!imp!Entity") != std::string::npos) {
-                    if (inEntityBlock) entitiesToCreate.push_back(current);
-                    inEntityBlock = true;
-                    current = PendingEntity{};
-                    section = Section::None;
-                    continue;
-                }
-            }
-
-            if (line == "Transform:") { section = Section::Transform; continue; }
-            if (line == "MeshRenderer3D:") { section = Section::MeshRenderer3D; continue; }
-
-            auto sep = line.find(':');
-            if (sep == std::string::npos) continue;
-            std::string key = trim(line.substr(0, sep));
-            std::string value = trim(line.substr(sep + 1));
-
-            if (key == "name") {
-                if (!value.empty()) current.tag = value;
-                continue;
-            }
-
-            auto parseObjVec3 = [&](const std::string& obj) -> glm::vec3 {
-                // Expect: {x: a, y: b, z: c} or {r: a, g: b, b: c}
-                glm::vec3 out{0.0f};
-                std::string s = obj;
-                if (!s.empty() && s.front() == '{' && s.back() == '}') {
-                    s = s.substr(1, s.size()-2);
-                }
-                std::stringstream ss2(s);
-                std::string part;
-                while (std::getline(ss2, part, ',')) {
-                    auto colon = part.find(':');
-                    if (colon == std::string::npos) continue;
-                    auto k = trim(part.substr(0, colon));
-                    auto v = trim(part.substr(colon+1));
-                    float f = std::stof(v);
-                    if (k == "x" || k == "r") out.x = f;
-                    else if (k == "y" || k == "g") out.y = f;
-                    else if (k == "z" || k == "b") out.z = f;
-                }
-                return out;
-            };
-
-            if (section == Section::Transform) {
-                if (key == "position") current.position = parseObjVec3(value);
-                else if (key == "rotation") current.rotation = parseObjVec3(value);
-                else if (key == "scale") current.scale = parseObjVec3(value);
-            } else if (section == Section::MeshRenderer3D) {
-                if (key == "modelPath") current.modelPath = value;
-                else if (key == "texturePath") current.texturePath = value;
-                else if (key == "color") current.color = parseObjVec3(value);
-            }
-        }
-        // Flush last block or default
-        if (inEntityBlock) {
-            entitiesToCreate.push_back(current);
-        } else {
-            // If no explicit entity block, use accumulated defaults
-            entitiesToCreate.push_back(current);
-        }
-    } else {
-        // No scene file; create one default entity
-        entitiesToCreate.push_back(PendingEntity{});
-    }
-
-    // Instantiate ECS entities and components
+    // Load GPU resources for all entities with MeshRenderer3D components
     auto& reg = ECSRegistry::getRegistry();
-    for (const auto& pe : entitiesToCreate) {
-        Entity e = reg.createEntity();
-        reg.addComponent<impgine::Transform>(e, { pe.position, pe.rotation, pe.scale });
-        if (pe.tag.has_value()) {
-            reg.addComponent<impgine::Tag>(e, impgine::Tag{ *pe.tag });
+    for (const auto& entity : reg.getEntities()) {
+        const auto& components = reg.getComponents(entity);
+        if (components.find(std::type_index(typeid(MeshRenderer3D))) != components.end()) {
+            auto& meshRenderer = reg.getComponent<MeshRenderer3D>(entity);
+            std::string modelPath = meshRenderer.mesh->modelPath;
+            std::string texturePath = meshRenderer.texture ? meshRenderer.texture->texturePath : "";
+
+            // Load mesh resources using ResourceManager
+            resourceManager->loadMeshResources(modelPath, texturePath);
         }
-
-        // Resolve paths relative to project
-        std::string fullModelPath = project.getFullPath(pe.modelPath);
-        std::string fullTexturePath = project.getFullPath(pe.texturePath);
-
-        reg.addComponent<impgine::MeshRenderer3D>(e, impgine::MeshRenderer3D{ std::make_shared<impgine::Mesh>(fullModelPath), std::make_shared<impgine::Texture2D>(fullTexturePath), pe.color });
-
-        // Load GPU resources for this entity's mesh
-        loadMeshResources(fullModelPath, fullTexturePath);
     }
 
-    createUniformBuffers();
-    createDescriptorSetLayout();
-    createDescriptorPool();
-    createDescriptorSets();
+    // Create uniform buffers and descriptor sets
+    uint32_t imageCount = static_cast<uint32_t>(swapChain->imageCount());
+    createUniformBuffers(device, physicalDevice, imageCount,
+                        uniformBuffers, uniformBuffersMemory);
+    createDescriptorSetLayout(device, descriptorSetLayout);
+
+    uint32_t meshCount = static_cast<uint32_t>(resourceManager->getMeshCache().size());
+    createDescriptorPool(device, meshCount, imageCount, descriptorPool);
+    createDescriptorSets(device, descriptorPool, descriptorSetLayout, uniformBuffers,
+                        imageCount, &resourceManager->getMeshCache());
+
     // Create pipeline layout now that descriptor set layout exists
     VkPushConstantRange pushConstantRange{};
     pushConstantRange.stageFlags = VK_SHADER_STAGE_VERTEX_BIT;
@@ -464,7 +306,7 @@ void Engine::loadScene(const std::string& scenePath) {
     std::string fragShader = project.getFullPath(project.shadersPath + "/frag.spv");
     pipeline = std::make_unique<Pipeline>(device, vertShader, fragShader, pipelineConfig);
 
-    createCommandBuffers();
+    createCommandBuffers(device, commandPool, SwapChain::MAX_FRAMES_IN_FLIGHT, commandBuffers);
 
     // Initialize UI renderer and a default window
     {
@@ -484,1172 +326,12 @@ void Engine::loadScene(const std::string& scenePath) {
     }
 }
 
-bool Engine::checkValidationLayerSupport() {
-    uint32_t layerCount;
-    vkEnumerateInstanceLayerProperties(&layerCount, nullptr);
-
-    std::vector<VkLayerProperties> availableLayers(layerCount);
-    vkEnumerateInstanceLayerProperties(&layerCount, availableLayers.data());
-
-    for (const char* layerName : validationLayers) {
-        bool layerFound = false;
-
-        for (const auto& layerProperties : availableLayers) {
-            if (strcmp(layerName, layerProperties.layerName) == 0) {
-                layerFound = true;
-                break;
-            }
-        }
-
-        if (!layerFound) {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-void Engine::populateDebugMessengerCreateInfo(VkDebugUtilsMessengerCreateInfoEXT& createInfo) {
-    createInfo = {};
-    createInfo.sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CREATE_INFO_EXT;
-    createInfo.messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT |
-                                 VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT |
-                                 VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT;
-    createInfo.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT |
-                             VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT |
-                             VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT;
-    createInfo.pfnUserCallback = debugCallback;
-}
-
-void Engine::setupDebugMessenger() {
-    if (!enableValidationLayers) return;
-
-    VkDebugUtilsMessengerCreateInfoEXT createInfo;
-    populateDebugMessengerCreateInfo(createInfo);
-
-    if (CreateDebugUtilsMessengerEXT(instance, &createInfo, nullptr, &debugMessenger) !=
-        VK_SUCCESS) {
-        throw std::runtime_error("failed to set up debug messenger!");
-    }
-}
-
-void Engine::createSurface() { window->createWindowSurface(instance, &surface); }
-
-bool Engine::isDeviceSuitable(VkPhysicalDevice device) {
-    QueueFamilyIndices indices = findQueueFamilies(device);
-
-    bool extensionsSupported = checkDeviceExtensionSupport(device);
-
-    bool swapChainAdequate = false;
-    if (extensionsSupported) {
-        auto swapChainSupport = SwapChain::querySwapChainSupport(device, surface);
-        swapChainAdequate =
-            !swapChainSupport.formats.empty() && !swapChainSupport.presentModes.empty();
-    }
-
-    VkPhysicalDeviceFeatures supportedFeatures;
-    vkGetPhysicalDeviceFeatures(device, &supportedFeatures);
-
-    return indices.isComplete() && extensionsSupported && swapChainAdequate && supportedFeatures.samplerAnisotropy;
-}
-
-bool Engine::checkDeviceExtensionSupport(VkPhysicalDevice device) {
-    uint32_t extensionCount;
-    vkEnumerateDeviceExtensionProperties(device, nullptr, &extensionCount, nullptr);
-
-    std::vector<VkExtensionProperties> availableExtensions(extensionCount);
-    vkEnumerateDeviceExtensionProperties(device, nullptr, &extensionCount,
-                                         availableExtensions.data());
-
-    std::set<std::string> requiredExtensions(deviceExtensions.begin(), deviceExtensions.end());
-
-    // Check if portability subset is available and add it if needed
-    bool portabilitySubsetAvailable = false;
-    for (const auto& extension : availableExtensions) {
-        if (strcmp(extension.extensionName, "VK_KHR_portability_subset") == 0) {
-            portabilitySubsetAvailable = true;
-        }
-        requiredExtensions.erase(extension.extensionName);
-    }
-
-    // Store whether we need portability subset for use in createLogicalDevice
-    needsPortabilitySubset = portabilitySubsetAvailable;
-
-    return requiredExtensions.empty();
-}
-
-void Engine::pickPhysicalDevice() {
-    uint32_t deviceCount = 0;
-    vkEnumeratePhysicalDevices(instance, &deviceCount, nullptr);
-    if (deviceCount == 0) {
-        throw std::runtime_error("failed to find GPUs with Vulkan support!");
-    }
-
-    std::vector<VkPhysicalDevice> devices(deviceCount);
-    vkEnumeratePhysicalDevices(instance, &deviceCount, devices.data());
-
-    for (const auto& device : devices) {
-        if (isDeviceSuitable(device)) {
-            physicalDevice = device;
-            msaaSamples = getMaxUsableSampleCount();
-            break;
-        }
-    }
-
-    if (physicalDevice == VK_NULL_HANDLE) {
-        throw std::runtime_error("failed to find a suitable GPU!");
-    }
-}
-
-QueueFamilyIndices Engine::findQueueFamilies(VkPhysicalDevice device) {
-    QueueFamilyIndices indices;
-
-    uint32_t queueFamilyCount = 0;
-    vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, nullptr);
-
-    std::vector<VkQueueFamilyProperties> queueFamilies(queueFamilyCount);
-    vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, queueFamilies.data());
-
-    int i = 0;
-    for (const auto& queueFamily : queueFamilies) {
-        if (queueFamily.queueFlags & VK_QUEUE_GRAPHICS_BIT) {
-            indices.graphicsFamily = i;
-        }
-
-        VkBool32 presentSupport = false;
-        vkGetPhysicalDeviceSurfaceSupportKHR(device, i, surface, &presentSupport);
-
-        if (presentSupport) {
-            indices.presentFamily = i;
-        }
-
-        if (indices.isComplete()) {
-            break;
-        }
-
-        i++;
-    }
-    return indices;
-}
-
-void Engine::createLogicalDevice() {
-    QueueFamilyIndices indices = findQueueFamilies(physicalDevice);
-
-    std::vector<VkDeviceQueueCreateInfo> queueCreateInfos;
-    std::set<uint32_t> uniqueQueueFamilies = {indices.graphicsFamily.value(),
-                                              indices.presentFamily.value()};
-
-    float queuePriority = 1.0f;
-    for (uint32_t queueFamily : uniqueQueueFamilies) {
-        VkDeviceQueueCreateInfo queueCreateInfo{};
-        queueCreateInfo.sType = VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO;
-        queueCreateInfo.queueFamilyIndex = queueFamily;
-        queueCreateInfo.queueCount = 1;
-        queueCreateInfo.pQueuePriorities = &queuePriority;
-        queueCreateInfos.push_back(queueCreateInfo);
-    }
-
-    VkPhysicalDeviceFeatures deviceFeatures{};
-    deviceFeatures.samplerAnisotropy = VK_TRUE;
-
-    VkDeviceCreateInfo createInfo{};
-    createInfo.sType = VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO;
-
-    createInfo.queueCreateInfoCount = static_cast<uint32_t>(queueCreateInfos.size());
-    createInfo.pQueueCreateInfos = queueCreateInfos.data();
-
-    createInfo.pEnabledFeatures = &deviceFeatures;
-
-    // Add portability subset extension if needed
-    std::vector<const char*> extensions = deviceExtensions;
-    if (needsPortabilitySubset) {
-        extensions.push_back("VK_KHR_portability_subset");
-    }
-
-    createInfo.enabledExtensionCount = static_cast<uint32_t>(extensions.size());
-    createInfo.ppEnabledExtensionNames = extensions.data();
-
-    if (enableValidationLayers) {
-        createInfo.enabledLayerCount = static_cast<uint32_t>(validationLayers.size());
-        createInfo.ppEnabledLayerNames = validationLayers.data();
-    } else {
-        createInfo.enabledLayerCount = 0;
-    }
-
-    if (vkCreateDevice(physicalDevice, &createInfo, nullptr, &device) != VK_SUCCESS) {
-        throw std::runtime_error("failed to create logical device!");
-    }
-
-    vkGetDeviceQueue(device, indices.graphicsFamily.value(), 0, &graphicsQueue);
-    vkGetDeviceQueue(device, indices.presentFamily.value(), 0, &presentQueue);
-}
-
-std::vector<const char*> Engine::getRequiredExtensions() {
-    uint32_t glfwExtensionCount = 0;
-    const char** glfwExtensions;
-    glfwExtensions = glfwGetRequiredInstanceExtensions(&glfwExtensionCount);
-
-    std::vector<const char*> extensions(glfwExtensions, glfwExtensions + glfwExtensionCount);
-
-    if (enableValidationLayers) {
-        extensions.push_back(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
-    }
-
-        // Required on MoltenVK/portability for VK_KHR_portability_subset dependency
-        extensions.push_back("VK_KHR_get_physical_device_properties2");
-
-    return extensions;
-}
-
-void Engine::createCommandPool() {
-    QueueFamilyIndices queueFamilyIndices = findQueueFamilies(physicalDevice);
-
-    VkCommandPoolCreateInfo poolInfo{};
-    poolInfo.sType = VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO;
-    poolInfo.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
-    poolInfo.queueFamilyIndex = queueFamilyIndices.graphicsFamily.value();
-
-    if (vkCreateCommandPool(device, &poolInfo, nullptr, &commandPool) != VK_SUCCESS) {
-        throw std::runtime_error("failed to create command pool!");
-    }
-}
-
-uint32_t Engine::findMemoryType(uint32_t typeFilter, VkMemoryPropertyFlags properties) {
-    VkPhysicalDeviceMemoryProperties memProperties;
-    vkGetPhysicalDeviceMemoryProperties(physicalDevice, &memProperties);
-
-    for (uint32_t i = 0; i < memProperties.memoryTypeCount; i++) {
-        if ((typeFilter & (1 << i)) && (memProperties.memoryTypes[i].propertyFlags & properties) == properties) {
-            return i;
-        }
-    }
-
-    throw std::runtime_error("failed to find suitable memory type!");
-}
-
-void Engine::createBuffer(VkDeviceSize size, VkBufferUsageFlags usage, VkMemoryPropertyFlags properties, VkBuffer& buffer, VkDeviceMemory& bufferMemory) {
-    VkBufferCreateInfo bufferInfo{};
-    bufferInfo.sType = VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO;
-    bufferInfo.size = size;
-    bufferInfo.usage = usage;
-    bufferInfo.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
-
-    if (vkCreateBuffer(device, &bufferInfo, nullptr, &buffer) != VK_SUCCESS) {
-        throw std::runtime_error("failed to create buffer!");
-    }
-
-    VkMemoryRequirements memRequirements;
-    vkGetBufferMemoryRequirements(device, buffer, &memRequirements);
-
-    VkMemoryAllocateInfo allocInfo{};
-    allocInfo.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-    allocInfo.allocationSize = memRequirements.size;
-    allocInfo.memoryTypeIndex = findMemoryType(memRequirements.memoryTypeBits, properties);
-
-    if (vkAllocateMemory(device, &allocInfo, nullptr, &bufferMemory) != VK_SUCCESS) {
-        throw std::runtime_error("failed to allocate buffer memory!");
-    }
-
-    vkBindBufferMemory(device, buffer, bufferMemory, 0);
-}
-
-void Engine::createImage(uint32_t width, uint32_t height, uint32_t mipLevels, VkSampleCountFlagBits numSamples, VkFormat format, VkImageTiling tiling, VkImageUsageFlags usage, VkMemoryPropertyFlags properties, VkImage& image, VkDeviceMemory& imageMemory) {
-    VkImageCreateInfo imageInfo{};
-    imageInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
-    imageInfo.imageType = VK_IMAGE_TYPE_2D;
-    imageInfo.extent.width = width;
-    imageInfo.extent.height = height;
-    imageInfo.extent.depth = 1;
-    imageInfo.mipLevels = mipLevels;
-    imageInfo.arrayLayers = 1;
-    imageInfo.format = format;
-    imageInfo.tiling = tiling;
-    imageInfo.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    imageInfo.usage = usage;
-    imageInfo.samples = numSamples;
-    imageInfo.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
-
-    if (vkCreateImage(device, &imageInfo, nullptr, &image) != VK_SUCCESS) {
-        throw std::runtime_error("echec de la creation d'une image!");
-    }
-
-    VkMemoryRequirements memRequirements;
-    vkGetImageMemoryRequirements(device, image, &memRequirements);
-
-    VkMemoryAllocateInfo allocInfo{};
-    allocInfo.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-    allocInfo.allocationSize = memRequirements.size;
-    allocInfo.memoryTypeIndex = findMemoryType(memRequirements.memoryTypeBits, properties);
-
-    if (vkAllocateMemory(device, &allocInfo, nullptr, &imageMemory) != VK_SUCCESS) {
-        throw std::runtime_error("echec de l'allocation de la memoire d'une image!");
-    }
-
-    vkBindImageMemory(device, image, imageMemory, 0);
-}
-
-VkImageView Engine::createImageView(VkImage image, VkFormat format, VkImageAspectFlags aspectFlags, uint32_t mipLevels) {
-    VkImageViewCreateInfo viewInfo{};
-    viewInfo.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
-    viewInfo.image = image;
-    viewInfo.viewType = VK_IMAGE_VIEW_TYPE_2D;
-    viewInfo.format = format;
-    viewInfo.subresourceRange.aspectMask = aspectFlags;
-    viewInfo.subresourceRange.baseMipLevel = 0;
-    viewInfo.subresourceRange.levelCount = mipLevels;
-    viewInfo.subresourceRange.baseArrayLayer = 0;
-    viewInfo.subresourceRange.layerCount = 1;
-
-    VkImageView imageView;
-    if (vkCreateImageView(device, &viewInfo, nullptr, &imageView) != VK_SUCCESS) {
-        throw std::runtime_error("échec de la creation de la vue sur une image!");
-    }
-
-    return imageView;
-}
-
-VkCommandBuffer Engine::beginSingleTimeCommands() {
-    VkCommandBufferAllocateInfo allocInfo{};
-    allocInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-    allocInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-    allocInfo.commandPool = commandPool;
-    allocInfo.commandBufferCount = 1;
-
-    VkCommandBuffer commandBuffer;
-    vkAllocateCommandBuffers(device, &allocInfo, &commandBuffer);
-
-    VkCommandBufferBeginInfo beginInfo{};
-    beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-    beginInfo.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT;
-
-    vkBeginCommandBuffer(commandBuffer, &beginInfo);
-
-    return commandBuffer;
-}
-
-void Engine::endSingleTimeCommands(VkCommandBuffer commandBuffer) {
-    vkEndCommandBuffer(commandBuffer);
-
-    VkSubmitInfo submitInfo{};
-    submitInfo.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
-    submitInfo.commandBufferCount = 1;
-    submitInfo.pCommandBuffers = &commandBuffer;
-
-    vkQueueSubmit(graphicsQueue, 1, &submitInfo, VK_NULL_HANDLE);
-    vkQueueWaitIdle(graphicsQueue);
-
-    vkFreeCommandBuffers(device, commandPool, 1, &commandBuffer);
-}
-
-void Engine::copyBuffer(VkBuffer srcBuffer, VkBuffer dstBuffer, VkDeviceSize size) {
-    VkCommandBuffer commandBuffer = beginSingleTimeCommands();
-
-    VkBufferCopy copyRegion{};
-    copyRegion.size = size;
-    vkCmdCopyBuffer(commandBuffer, srcBuffer, dstBuffer, 1, &copyRegion);
-
-    endSingleTimeCommands(commandBuffer);
-}
-
-void Engine::transitionImageLayout(VkImage image, VkFormat format, VkImageLayout oldLayout, VkImageLayout newLayout, uint32_t mipLevels) {
-    VkCommandBuffer commandBuffer = beginSingleTimeCommands();
-
-    VkImageMemoryBarrier barrier{};
-    barrier.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER;
-    barrier.oldLayout = oldLayout;
-    barrier.newLayout = newLayout;
-    barrier.srcQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-    barrier.dstQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-    barrier.image = image;
-    
-    if (newLayout == VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL) {
-        barrier.subresourceRange.aspectMask = VK_IMAGE_ASPECT_DEPTH_BIT;
-        
-        if (hasStencilComponent(format)) {
-            barrier.subresourceRange.aspectMask |= VK_IMAGE_ASPECT_STENCIL_BIT;
-        }
-    } else {
-    barrier.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    }
-    
-    barrier.subresourceRange.baseMipLevel = 0;
-    barrier.subresourceRange.levelCount = mipLevels;
-    barrier.subresourceRange.baseArrayLayer = 0;
-    barrier.subresourceRange.layerCount = 1;
-
-    VkPipelineStageFlags sourceStage;
-    VkPipelineStageFlags destinationStage;
-
-    if (oldLayout == VK_IMAGE_LAYOUT_UNDEFINED && newLayout == VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL) {
-        barrier.srcAccessMask = 0;
-        barrier.dstAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT;
-
-        sourceStage = VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT;
-        destinationStage = VK_PIPELINE_STAGE_TRANSFER_BIT;
-    } else if (oldLayout == VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL && newLayout == VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL) {
-        barrier.srcAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT;
-        barrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT;
-
-        sourceStage = VK_PIPELINE_STAGE_TRANSFER_BIT;
-        destinationStage = VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT;
-    } else if (oldLayout == VK_IMAGE_LAYOUT_UNDEFINED && newLayout == VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL) {
-        barrier.srcAccessMask = 0;
-        barrier.dstAccessMask = VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_READ_BIT | VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT;
-
-        sourceStage = VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT;
-        destinationStage = VK_PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT;
-    } else {
-        throw std::invalid_argument("transition d'organisation non supportée!");
-    }
-
-    vkCmdPipelineBarrier(
-        commandBuffer,
-        sourceStage, destinationStage,
-        0,
-        0, nullptr,
-        0, nullptr,
-        1, &barrier
-    );
-
-    endSingleTimeCommands(commandBuffer);
-}
-
-void Engine::copyBufferToImage(VkBuffer buffer, VkImage image, uint32_t width, uint32_t height) {
-    VkCommandBuffer commandBuffer = beginSingleTimeCommands();
-
-    VkBufferImageCopy region{};
-    region.bufferOffset = 0;
-    region.bufferRowLength = 0;
-    region.bufferImageHeight = 0;
-
-    region.imageSubresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    region.imageSubresource.mipLevel = 0;
-    region.imageSubresource.baseArrayLayer = 0;
-    region.imageSubresource.layerCount = 1;
-
-    region.imageOffset = {0, 0, 0};
-    region.imageExtent = {
-        width,
-        height,
-        1
-    };
-
-    vkCmdCopyBufferToImage(
-        commandBuffer,
-        buffer,
-        image,
-        VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL,
-        1,
-        &region
-    );
-
-    endSingleTimeCommands(commandBuffer);
-}
-
-void Engine::loadModel(const std::string& path, std::vector<Vertex>& vertices, std::vector<uint32_t>& indices) {
-    tinyobj::attrib_t attrib;
-    std::vector<tinyobj::shape_t> shapes;
-    std::vector<tinyobj::material_t> materials;
-    std::string warn, err;
-
-    if (!tinyobj::LoadObj(&attrib, &shapes, &materials, &warn, &err, path.c_str())) {
-        throw std::runtime_error(err);
-    }
-
-    std::unordered_map<Vertex, uint32_t> uniqueVertices{};
-
-    for (const auto& shape : shapes) {
-        for (const auto& index : shape.mesh.indices) {
-            Vertex vertex{};
-
-            vertex.pos = {
-                attrib.vertices[3 * index.vertex_index + 0],
-                attrib.vertices[3 * index.vertex_index + 1],
-                attrib.vertices[3 * index.vertex_index + 2]
-            };
-
-            if (index.texcoord_index >= 0 && static_cast<size_t>(index.texcoord_index) < attrib.texcoords.size() / 2) {
-                vertex.texCoord = {
-                    attrib.texcoords[2 * index.texcoord_index + 0],
-                    1.0f - attrib.texcoords[2 * index.texcoord_index + 1]
-                };
-            } else {
-                vertex.texCoord = {0.0f, 0.0f};
-            }
-
-            vertex.color = {1.0f, 1.0f, 1.0f};
-
-            if (uniqueVertices.count(vertex) == 0) {
-                uniqueVertices[vertex] = static_cast<uint32_t>(vertices.size());
-                vertices.push_back(vertex);
-            }
-
-            indices.push_back(uniqueVertices[vertex]);
-        }
-    }
-}
-
-void Engine::createVertexBuffer(const std::vector<Vertex>& vertices, VkBuffer& buffer, VkDeviceMemory& bufferMemory) {
-    VkDeviceSize bufferSize = sizeof(vertices[0]) * vertices.size();
-
-    VkBuffer stagingBuffer;
-    VkDeviceMemory stagingBufferMemory;
-    createBuffer(bufferSize, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT, stagingBuffer, stagingBufferMemory);
-
-    void* data;
-    vkMapMemory(device, stagingBufferMemory, 0, bufferSize, 0, &data);
-        memcpy(data, vertices.data(), (size_t) bufferSize);
-    vkUnmapMemory(device, stagingBufferMemory);
-
-    createBuffer(bufferSize, VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_VERTEX_BUFFER_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, buffer, bufferMemory);
-
-    copyBuffer(stagingBuffer, buffer, bufferSize);
-
-    vkDestroyBuffer(device, stagingBuffer, nullptr);
-    vkFreeMemory(device, stagingBufferMemory, nullptr);
-}
-
-void Engine::createIndexBuffer(const std::vector<uint32_t>& indices, VkBuffer& buffer, VkDeviceMemory& bufferMemory) {
-    VkDeviceSize bufferSize = sizeof(indices[0]) * indices.size();
-
-    VkBuffer stagingBuffer;
-    VkDeviceMemory stagingBufferMemory;
-    createBuffer(bufferSize, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT, stagingBuffer, stagingBufferMemory);
-
-    void* data;
-    vkMapMemory(device, stagingBufferMemory, 0, bufferSize, 0, &data);
-    memcpy(data, indices.data(), (size_t) bufferSize);
-    vkUnmapMemory(device, stagingBufferMemory);
-
-    createBuffer(bufferSize, VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_INDEX_BUFFER_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, buffer, bufferMemory);
-
-    copyBuffer(stagingBuffer, buffer, bufferSize);
-
-    vkDestroyBuffer(device, stagingBuffer, nullptr);
-    vkFreeMemory(device, stagingBufferMemory, nullptr);
-}
-
-void Engine::createUniformBuffers() {
-    VkDeviceSize bufferSize = sizeof(UniformBufferObject);
-
-    uniformBuffers.resize(swapChain->imageCount());
-    uniformBuffersMemory.resize(swapChain->imageCount());
-
-    for (size_t i = 0; i < swapChain->imageCount(); i++) {
-        createBuffer(
-            bufferSize,
-            VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT,
-            VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT,
-            uniformBuffers[i], uniformBuffersMemory[i]
-        );
-    }
-}
-
-void Engine::createDescriptorSetLayout() {
-    VkDescriptorSetLayoutBinding uboLayoutBinding{};
-    uboLayoutBinding.binding = 0;
-    uboLayoutBinding.descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-    uboLayoutBinding.descriptorCount = 1;
-    uboLayoutBinding.stageFlags = VK_SHADER_STAGE_VERTEX_BIT;
-    uboLayoutBinding.pImmutableSamplers = nullptr; // Optional
-
-    VkDescriptorSetLayoutBinding samplerLayoutBinding{};
-    samplerLayoutBinding.binding = 1;
-    samplerLayoutBinding.descriptorCount = 1;
-    samplerLayoutBinding.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    samplerLayoutBinding.pImmutableSamplers = nullptr;
-    samplerLayoutBinding.stageFlags = VK_SHADER_STAGE_FRAGMENT_BIT;
-
-    std::array<VkDescriptorSetLayoutBinding, 2> bindings = {uboLayoutBinding, samplerLayoutBinding};
-    VkDescriptorSetLayoutCreateInfo layoutInfo{};
-    layoutInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-    layoutInfo.bindingCount = static_cast<uint32_t>(bindings.size());
-    layoutInfo.pBindings = bindings.data();
-
-    if (vkCreateDescriptorSetLayout(device, &layoutInfo, nullptr, &descriptorSetLayout) != VK_SUCCESS) {
-        throw std::runtime_error("failed to create descriptor set layout!");
-    }
-}
-
-void Engine::createDescriptorPool() {
-    // Calculate total descriptor sets needed: one per mesh per frame
-    uint32_t maxSets = static_cast<uint32_t>(meshCache.size() * swapChain->imageCount());
-
-    std::array<VkDescriptorPoolSize, 2> poolSizes{};
-    poolSizes[0].type = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-    poolSizes[0].descriptorCount = maxSets;
-    poolSizes[1].type = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    poolSizes[1].descriptorCount = maxSets;
-
-    VkDescriptorPoolCreateInfo poolInfo{};
-    poolInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-    poolInfo.poolSizeCount = static_cast<uint32_t>(poolSizes.size());
-    poolInfo.pPoolSizes = poolSizes.data();
-    poolInfo.maxSets = maxSets;
-
-    if (vkCreateDescriptorPool(device, &poolInfo, nullptr, &descriptorPool) != VK_SUCCESS) {
-        throw std::runtime_error("failed to create descriptor pool!");
-    }
-}
-
-void Engine::createDescriptorSets() {
-    if (meshCache.empty()) {
-        throw std::runtime_error("No meshes loaded before creating descriptor sets!");
-    }
-
-    // Create descriptor sets for each mesh
-    for (auto& [modelPath, meshResources] : meshCache) {
-        std::vector<VkDescriptorSetLayout> layouts(swapChain->imageCount(), descriptorSetLayout);
-        VkDescriptorSetAllocateInfo allocInfo{};
-        allocInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-        allocInfo.descriptorPool = descriptorPool;
-        allocInfo.descriptorSetCount = static_cast<uint32_t>(swapChain->imageCount());
-        allocInfo.pSetLayouts = layouts.data();
-
-        meshResources.descriptorSets.resize(swapChain->imageCount());
-        if (vkAllocateDescriptorSets(device, &allocInfo, meshResources.descriptorSets.data()) != VK_SUCCESS) {
-            throw std::runtime_error("failed to allocate descriptor sets!");
-        }
-
-        // Update descriptor sets for each frame with this mesh's texture
-        for (size_t i = 0; i < swapChain->imageCount(); i++) {
-            VkDescriptorBufferInfo bufferInfo{};
-            bufferInfo.buffer = uniformBuffers[i];
-            bufferInfo.offset = 0;
-            bufferInfo.range = sizeof(UniformBufferObject);
-
-            VkDescriptorImageInfo imageInfo{};
-            imageInfo.imageLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-            imageInfo.imageView = meshResources.textureImageView;
-            imageInfo.sampler = meshResources.textureSampler;
-
-            std::array<VkWriteDescriptorSet, 2> descriptorWrites{};
-
-            descriptorWrites[0].sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-            descriptorWrites[0].dstSet = meshResources.descriptorSets[i];
-            descriptorWrites[0].dstBinding = 0;
-            descriptorWrites[0].dstArrayElement = 0;
-            descriptorWrites[0].descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-            descriptorWrites[0].descriptorCount = 1;
-            descriptorWrites[0].pBufferInfo = &bufferInfo;
-
-            descriptorWrites[1].sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-            descriptorWrites[1].dstSet = meshResources.descriptorSets[i];
-            descriptorWrites[1].dstBinding = 1;
-            descriptorWrites[1].dstArrayElement = 0;
-            descriptorWrites[1].descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-            descriptorWrites[1].descriptorCount = 1;
-            descriptorWrites[1].pImageInfo = &imageInfo;
-
-            vkUpdateDescriptorSets(device, static_cast<uint32_t>(descriptorWrites.size()), descriptorWrites.data(), 0, nullptr);
-        }
-    }
-}
-
-void Engine::createTextureImage(const std::string& texturePath, VkImage& image, VkDeviceMemory& imageMemory, uint32_t& mipLevels) {
-    int texWidth, texHeight, texChannels;
-    stbi_uc* pixels = stbi_load(texturePath.c_str(), &texWidth, &texHeight, &texChannels, STBI_rgb_alpha);
-
-    if (!pixels) {
-        std::cout << "Texture not found in textures folder, trying the base texture.jpg" << std::endl;
-        // callback to the texture.jpg base image
-        pixels = stbi_load("textures/texture.jpg", &texWidth, &texHeight, &texChannels, STBI_rgb_alpha);
-
-        if (!pixels) {
-            throw std::runtime_error("échec du chargement d'une image!");
-        }
-    }
-
-    // Calculate mipLevels and imageSize AFTER we have valid texture dimensions
-    mipLevels = static_cast<uint32_t>(std::floor(std::log2(std::max(texWidth, texHeight)))) + 1;
-    VkDeviceSize imageSize = texWidth * texHeight * 4;
-
-    VkBuffer stagingBuffer;
-    VkDeviceMemory stagingBufferMemory;
-    createBuffer(imageSize, VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT, stagingBuffer, stagingBufferMemory);
-
-    void* data;
-    vkMapMemory(device, stagingBufferMemory, 0, imageSize, 0, &data);
-        memcpy(data, pixels, static_cast<size_t>(imageSize));
-    vkUnmapMemory(device, stagingBufferMemory);
-
-    stbi_image_free(pixels);
-
-    createImage(texWidth, texHeight, mipLevels, VK_SAMPLE_COUNT_1_BIT, VK_FORMAT_R8G8B8A8_SRGB, VK_IMAGE_TILING_OPTIMAL, VK_IMAGE_USAGE_TRANSFER_SRC_BIT | VK_IMAGE_USAGE_TRANSFER_DST_BIT | VK_IMAGE_USAGE_SAMPLED_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, image, imageMemory);
-
-    transitionImageLayout(image, VK_FORMAT_R8G8B8A8_SRGB, VK_IMAGE_LAYOUT_UNDEFINED, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL, mipLevels);
-    copyBufferToImage(stagingBuffer, image, static_cast<uint32_t>(texWidth), static_cast<uint32_t>(texHeight));
-    //transitionné vers VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL lors de la generation des mipmaps
-
-    generateMipmaps(image, VK_FORMAT_R8G8B8A8_SRGB, texWidth, texHeight, mipLevels);
-
-    vkDestroyBuffer(device, stagingBuffer, nullptr);
-    vkFreeMemory(device, stagingBufferMemory, nullptr);
-}
-
-VkImageView Engine::createTextureImageView(VkImage image, uint32_t mipLevels) {
-    return createImageView(image, VK_FORMAT_R8G8B8A8_SRGB, VK_IMAGE_ASPECT_COLOR_BIT, mipLevels);
-}
-
-VkSampler Engine::createTextureSampler(uint32_t mipLevels) {
-    VkSamplerCreateInfo samplerInfo{};
-    samplerInfo.sType = VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO;
-    samplerInfo.magFilter = VK_FILTER_LINEAR;
-    samplerInfo.minFilter = VK_FILTER_LINEAR;
-    samplerInfo.addressModeU = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-    samplerInfo.addressModeV = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-    samplerInfo.addressModeW = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-    samplerInfo.anisotropyEnable = VK_TRUE;
-    samplerInfo.maxAnisotropy = 16.0f;
-    samplerInfo.borderColor = VK_BORDER_COLOR_INT_OPAQUE_BLACK;
-    samplerInfo.unnormalizedCoordinates = VK_FALSE;
-    samplerInfo.compareEnable = VK_FALSE;
-    samplerInfo.compareOp = VK_COMPARE_OP_ALWAYS;
-    samplerInfo.mipmapMode = VK_SAMPLER_MIPMAP_MODE_LINEAR;
-    samplerInfo.minLod = 0.0f;
-    samplerInfo.maxLod = static_cast<float>(mipLevels);
-    samplerInfo.mipLodBias = 0.0f;
-
-    VkSampler sampler;
-    if (vkCreateSampler(device, &samplerInfo, nullptr, &sampler) != VK_SUCCESS) {
-        throw std::runtime_error("échec de la creation d'un sampler!");
-    }
-    return sampler;
-}
-
-MeshGPUResources& Engine::loadMeshResources(const std::string& modelPath, const std::string& texturePath) {
-    // Check if already loaded
-    auto it = meshCache.find(modelPath);
-    if (it != meshCache.end()) {
-        return it->second;
-    }
-
-    // Load new mesh
-    MeshGPUResources resources;
-    loadModel(modelPath, resources.vertices, resources.indices);
-    createVertexBuffer(resources.vertices, resources.vertexBuffer, resources.vertexBufferMemory);
-    createIndexBuffer(resources.indices, resources.indexBuffer, resources.indexBufferMemory);
-    createTextureImage(texturePath, resources.textureImage, resources.textureImageMemory, resources.mipLevels);
-    resources.textureImageView = createTextureImageView(resources.textureImage, resources.mipLevels);
-    resources.textureSampler = createTextureSampler(resources.mipLevels);
-
-    meshCache[modelPath] = resources;
-    return meshCache[modelPath];
-}
-
-VkFormat Engine::findSupportedFormat(const std::vector<VkFormat>& candidates, VkImageTiling tiling, VkFormatFeatureFlags features) {
-    for (VkFormat format : candidates) {
-        VkFormatProperties props;
-        vkGetPhysicalDeviceFormatProperties(physicalDevice, format, &props);
-
-        if (tiling == VK_IMAGE_TILING_LINEAR && (props.linearTilingFeatures & features) == features) {
-            return format;
-        } else if (tiling == VK_IMAGE_TILING_OPTIMAL && (props.optimalTilingFeatures & features) == features) {
-            return format;
-        }
-    }
-
-    throw std::runtime_error("failed to find supported format!");
-}
-
-VkFormat Engine::findDepthFormat() {
-    return findSupportedFormat(
-        {VK_FORMAT_D32_SFLOAT, VK_FORMAT_D32_SFLOAT_S8_UINT, VK_FORMAT_D24_UNORM_S8_UINT},
-        VK_IMAGE_TILING_OPTIMAL,
-        VK_FORMAT_FEATURE_DEPTH_STENCIL_ATTACHMENT_BIT
-    );
-}
-
-bool Engine::hasStencilComponent(VkFormat format) {
-    return format == VK_FORMAT_D32_SFLOAT_S8_UINT || format == VK_FORMAT_D24_UNORM_S8_UINT;
-}
-
-VkSampleCountFlagBits Engine::getMaxUsableSampleCount() {
-    VkPhysicalDeviceProperties physicalDeviceProperties;
-    vkGetPhysicalDeviceProperties(physicalDevice, &physicalDeviceProperties);
-
-    VkSampleCountFlags counts = physicalDeviceProperties.limits.framebufferColorSampleCounts & physicalDeviceProperties.limits.framebufferDepthSampleCounts;
-    if (counts & VK_SAMPLE_COUNT_64_BIT) { return VK_SAMPLE_COUNT_64_BIT; }
-    if (counts & VK_SAMPLE_COUNT_32_BIT) { return VK_SAMPLE_COUNT_32_BIT; }
-    if (counts & VK_SAMPLE_COUNT_16_BIT) { return VK_SAMPLE_COUNT_16_BIT; }
-    if (counts & VK_SAMPLE_COUNT_8_BIT) { return VK_SAMPLE_COUNT_8_BIT; }
-    if (counts & VK_SAMPLE_COUNT_4_BIT) { return VK_SAMPLE_COUNT_4_BIT; }
-    if (counts & VK_SAMPLE_COUNT_2_BIT) { return VK_SAMPLE_COUNT_2_BIT; }
-
-    return VK_SAMPLE_COUNT_1_BIT;
-}
-
-void Engine::createColorResources() {
-    VkFormat colorFormat = swapChain->getSwapChainImageFormat();
-
-    createImage(swapChain->getSwapChainExtent().width, swapChain->getSwapChainExtent().height, 1, msaaSamples, colorFormat, VK_IMAGE_TILING_OPTIMAL, VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT | VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, colorImage, colorImageMemory);
-    colorImageView = createImageView(colorImage, colorFormat, VK_IMAGE_ASPECT_COLOR_BIT, 1);
-}
-
-void Engine::createDepthResources() {
-    VkFormat depthFormat = findDepthFormat();
-    
-    createImage(swapChain->getSwapChainExtent().width, swapChain->getSwapChainExtent().height, 1, msaaSamples, depthFormat, VK_IMAGE_TILING_OPTIMAL, VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, depthImage, depthImageMemory);
-    depthImageView = createImageView(depthImage, depthFormat, VK_IMAGE_ASPECT_DEPTH_BIT, 1);
-}
-
-void Engine::createRenderPass() {
-    VkAttachmentDescription colorAttachment{};
-    colorAttachment.format = swapChain->getSwapChainImageFormat();
-    colorAttachment.samples = msaaSamples;
-    colorAttachment.loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-    colorAttachment.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-    colorAttachment.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-    colorAttachment.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-    colorAttachment.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    colorAttachment.finalLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-    VkAttachmentDescription depthAttachment{};
-    depthAttachment.format = findDepthFormat();
-    depthAttachment.samples = msaaSamples;
-    depthAttachment.loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-    depthAttachment.storeOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-    depthAttachment.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-    depthAttachment.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-    depthAttachment.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    depthAttachment.finalLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-
-    VkAttachmentDescription colorAttachmentResolve{};
-    colorAttachmentResolve.format = swapChain->getSwapChainImageFormat();
-    colorAttachmentResolve.samples = VK_SAMPLE_COUNT_1_BIT;
-    colorAttachmentResolve.loadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-    colorAttachmentResolve.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-    colorAttachmentResolve.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-    colorAttachmentResolve.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-    colorAttachmentResolve.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    colorAttachmentResolve.finalLayout = VK_IMAGE_LAYOUT_PRESENT_SRC_KHR;
-
-    VkAttachmentReference colorAttachmentRef{};
-    colorAttachmentRef.attachment = 0;
-    colorAttachmentRef.layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-    VkAttachmentReference depthAttachmentRef{};
-    depthAttachmentRef.attachment = 1;
-    depthAttachmentRef.layout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-
-    VkAttachmentReference colorAttachmentResolveRef{};
-    colorAttachmentResolveRef.attachment = 2;
-    colorAttachmentResolveRef.layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-    VkSubpassDescription subpass{};
-    subpass.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-    subpass.colorAttachmentCount = 1;
-    subpass.pColorAttachments = &colorAttachmentRef;
-    subpass.pDepthStencilAttachment = &depthAttachmentRef;
-    subpass.pResolveAttachments = &colorAttachmentResolveRef;
-
-    VkSubpassDependency dependency{};
-    dependency.srcSubpass = VK_SUBPASS_EXTERNAL;
-    dependency.dstSubpass = 0;
-    dependency.srcStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT | VK_PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT;
-    dependency.srcAccessMask = 0;
-    dependency.dstStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT | VK_PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT;
-    dependency.dstAccessMask = VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT | VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT;
-
-    std::array<VkAttachmentDescription, 3> attachments = {colorAttachment, depthAttachment, colorAttachmentResolve};
-
-    VkRenderPassCreateInfo renderPassInfo{};
-    renderPassInfo.sType = VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO;
-    renderPassInfo.attachmentCount = static_cast<uint32_t>(attachments.size());
-    renderPassInfo.pAttachments = attachments.data();
-    renderPassInfo.subpassCount = 1;
-    renderPassInfo.pSubpasses = &subpass;
-    renderPassInfo.dependencyCount = 1;
-    renderPassInfo.pDependencies = &dependency;
-
-    if (vkCreateRenderPass(device, &renderPassInfo, nullptr, &renderPass) != VK_SUCCESS) {
-        throw std::runtime_error("failed to create render pass!");
-    }
-}
-
-void Engine::createFramebuffers() {
-    swapChainFramebuffers.resize(swapChain->imageCount());
-
-    for (size_t i = 0; i < swapChain->imageCount(); i++) {
-        std::array<VkImageView, 3> attachments = {
-            colorImageView,
-            depthImageView,
-            swapChain->getImageView(i)
-        };
-
-        VkFramebufferCreateInfo framebufferInfo{};
-        framebufferInfo.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
-        framebufferInfo.renderPass = renderPass;
-        framebufferInfo.attachmentCount = static_cast<uint32_t>(attachments.size());
-        framebufferInfo.pAttachments = attachments.data();
-        framebufferInfo.width = swapChain->getSwapChainExtent().width;
-        framebufferInfo.height = swapChain->getSwapChainExtent().height;
-        framebufferInfo.layers = 1;
-
-        if (vkCreateFramebuffer(device, &framebufferInfo, nullptr, &swapChainFramebuffers[i]) != VK_SUCCESS) {
-            throw std::runtime_error("failed to create framebuffer!");
-        }
-    }
-}
-
-void Engine::generateMipmaps(VkImage image, VkFormat imageFormat, int32_t texWidth, int32_t texHeight, uint32_t mipLevels) {
-    // Vérifions si l'image supporte le filtrage linéaire
-    VkFormatProperties formatProperties;
-    vkGetPhysicalDeviceFormatProperties(physicalDevice, imageFormat, &formatProperties);
-
-    if (!(formatProperties.optimalTilingFeatures & VK_FORMAT_FEATURE_SAMPLED_IMAGE_FILTER_LINEAR_BIT)) {
-        throw std::runtime_error("le format de l'image texture ne supporte pas le filtrage lineaire!");
-    }
-
-    VkCommandBuffer commandBuffer = beginSingleTimeCommands();
-
-    VkImageMemoryBarrier barrier{};
-    barrier.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER;
-    barrier.image = image;
-    barrier.srcQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-    barrier.dstQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-    barrier.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    barrier.subresourceRange.baseArrayLayer = 0;
-    barrier.subresourceRange.layerCount = 1;
-    barrier.subresourceRange.levelCount = 1;
-
-    int32_t mipWidth = texWidth;
-    int32_t mipHeight = texHeight;
-
-    for (uint32_t i = 1; i < mipLevels; i++) {
-        barrier.subresourceRange.baseMipLevel = i - 1;
-        barrier.oldLayout = VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL;
-        barrier.newLayout = VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL;
-        barrier.srcAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT;
-        barrier.dstAccessMask = VK_ACCESS_TRANSFER_READ_BIT;
-
-        vkCmdPipelineBarrier(commandBuffer,
-            VK_PIPELINE_STAGE_TRANSFER_BIT, VK_PIPELINE_STAGE_TRANSFER_BIT, 0,
-            0, nullptr,
-            0, nullptr,
-            1, &barrier);
-
-        VkImageBlit blit{};
-        blit.srcOffsets[0] = { 0, 0, 0 };
-        blit.srcOffsets[1] = { mipWidth, mipHeight, 1 };
-        blit.srcSubresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-        blit.srcSubresource.mipLevel = i - 1;
-        blit.srcSubresource.baseArrayLayer = 0;
-        blit.srcSubresource.layerCount = 1;
-        blit.dstOffsets[0] = { 0, 0, 0 };
-        blit.dstOffsets[1] = { mipWidth > 1 ? mipWidth / 2 : 1, mipHeight > 1 ? mipHeight / 2 : 1, 1 };
-        blit.dstSubresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-        blit.dstSubresource.mipLevel = i;
-        blit.dstSubresource.baseArrayLayer = 0;
-        blit.dstSubresource.layerCount = 1;
-
-        vkCmdBlitImage(commandBuffer,
-            image, VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL,
-            image, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL,
-            1, &blit,
-            VK_FILTER_LINEAR);
-
-        barrier.oldLayout = VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL;
-        barrier.newLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-        barrier.srcAccessMask = VK_ACCESS_TRANSFER_READ_BIT;
-        barrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT;
-
-        vkCmdPipelineBarrier(commandBuffer,
-            VK_PIPELINE_STAGE_TRANSFER_BIT, VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT, 0,
-            0, nullptr,
-            0, nullptr,
-            1, &barrier);
-
-        if (mipWidth > 1) mipWidth /= 2;
-        if (mipHeight > 1) mipHeight /= 2;
-    }
-
-    barrier.subresourceRange.baseMipLevel = mipLevels - 1;
-    barrier.oldLayout = VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL;
-    barrier.newLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-    barrier.srcAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT;
-    barrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT;
-
-    vkCmdPipelineBarrier(commandBuffer,
-        VK_PIPELINE_STAGE_TRANSFER_BIT, VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT, 0,
-        0, nullptr,
-        0, nullptr,
-        1, &barrier);
-
-    endSingleTimeCommands(commandBuffer);
-}
-
-void Engine::updateUniformBuffer(uint32_t currentImage) {
-    auto currentTime = std::chrono::high_resolution_clock::now();
-    (void)currentTime; // silence unused for now
-
-    // Set up projection matrix (only needs to be set once unless aspect ratio changes)
-    camera.setPerspectiveProjection(glm::radians(45.0f), swapChain->getSwapChainExtent().width / (float) swapChain->getSwapChainExtent().height, 0.1f, 100.0f);
-    
-    // Update view matrix based on current camera state
-    camera.updateViewMatrix();
-
-    UniformBufferObject ubo{};
-    ubo.view = camera.getView();
-    ubo.proj = camera.getProjection();
-
-    void* data;
-    vkMapMemory(device, uniformBuffersMemory[currentImage], 0, sizeof(ubo), 0, &data);
-        memcpy(data, &ubo, sizeof(ubo));
-    vkUnmapMemory(device, uniformBuffersMemory[currentImage]);
-}
-
-void Engine::createCommandBuffers() {
-    commandBuffers.resize(SwapChain::MAX_FRAMES_IN_FLIGHT);
-
-    VkCommandBufferAllocateInfo allocInfo{};
-    allocInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-    allocInfo.commandPool = commandPool;
-    allocInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-    allocInfo.commandBufferCount = (uint32_t)commandBuffers.size();
-
-    if (vkAllocateCommandBuffers(device, &allocInfo, commandBuffers.data()) != VK_SUCCESS) {
-        throw std::runtime_error("failed to allocate command buffers!");
-    }
-}
-
-void Engine::recordCommandBuffer(VkCommandBuffer commandBuffer, uint32_t imageIndex) {
-    VkCommandBufferBeginInfo beginInfo{};
-    beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-
-    if (vkBeginCommandBuffer(commandBuffer, &beginInfo) != VK_SUCCESS) {
-        throw std::runtime_error("failed to begin recording command buffer!");
-    }
-
-    // Add memory barrier to ensure previous frame is complete
-    VkMemoryBarrier memoryBarrier{};
-    memoryBarrier.sType = VK_STRUCTURE_TYPE_MEMORY_BARRIER;
-    memoryBarrier.srcAccessMask = VK_ACCESS_MEMORY_WRITE_BIT;
-    memoryBarrier.dstAccessMask = VK_ACCESS_MEMORY_READ_BIT | VK_ACCESS_MEMORY_WRITE_BIT;
-    
-    vkCmdPipelineBarrier(
-        commandBuffer,
-        VK_PIPELINE_STAGE_ALL_COMMANDS_BIT,
-        VK_PIPELINE_STAGE_ALL_COMMANDS_BIT,
-        0,
-        1, &memoryBarrier,
-        0, nullptr,
-        0, nullptr
-    );
-
-    VkRenderPassBeginInfo renderPassInfo{};
-    renderPassInfo.sType = VK_STRUCTURE_TYPE_RENDER_PASS_BEGIN_INFO;
-    renderPassInfo.renderPass = renderPass;
-    renderPassInfo.framebuffer = swapChainFramebuffers[imageIndex];
-    renderPassInfo.renderArea.offset = {0, 0};
-    renderPassInfo.renderArea.extent = swapChain->getSwapChainExtent();
-
-    std::array<VkClearValue, 3> clearValues{};
-    clearValues[0].color = {{0.01f, 0.01f, 0.01f, 1.0f}};  // Color attachment (MSAA)
-    clearValues[1].depthStencil = {1.0f, 0};                // Depth attachment
-    clearValues[2].color = {{0.01f, 0.01f, 0.01f, 1.0f}};  // Resolve attachment
-    renderPassInfo.clearValueCount = static_cast<uint32_t>(clearValues.size());
-    renderPassInfo.pClearValues = clearValues.data();
-
-    vkCmdBeginRenderPass(commandBuffer, &renderPassInfo, VK_SUBPASS_CONTENTS_INLINE);
-
-    // Set viewport/scissor to UI center viewport so 3D scene renders only in middle
-    if (uiRenderer) {
-        // Ensure layout is up to date before using it (use framebuffer size for DPI correctness)
-        int ww = 0, wh = 0;
-        window->getFramebufferSize(&ww, &wh);
-        uiRenderer->getLayout().setWindowSize(ww, wh);
-        uiRenderer->getLayout().computeLayout(uiWindows);
-
-        glm::vec4 rect = uiRenderer->getLayout().getViewportRect();
-        VkViewport viewport{};
-        viewport.x = rect.x;
-        viewport.y = rect.y;
-        viewport.width = rect.z;
-        viewport.height = rect.w;
-        viewport.minDepth = 0.0f;
-        viewport.maxDepth = 1.0f;
-        vkCmdSetViewport(commandBuffer, 0, 1, &viewport);
-
-        VkRect2D scissor{};
-        scissor.offset = { static_cast<int32_t>(rect.x), static_cast<int32_t>(rect.y) };
-        scissor.extent = { static_cast<uint32_t>(rect.z), static_cast<uint32_t>(rect.w) };
-        vkCmdSetScissor(commandBuffer, 0, 1, &scissor);
-    } else {
-        VkViewport viewport{};
-        viewport.x = 0.0f;
-        viewport.y = 0.0f;
-        viewport.width = static_cast<float>(swapChain->getSwapChainExtent().width);
-        viewport.height = static_cast<float>(swapChain->getSwapChainExtent().height);
-        viewport.minDepth = 0.0f;
-        viewport.maxDepth = 1.0f;
-        vkCmdSetViewport(commandBuffer, 0, 1, &viewport);
-
-        VkRect2D scissor{};
-        scissor.offset = {0, 0};
-        scissor.extent = swapChain->getSwapChainExtent();
-        vkCmdSetScissor(commandBuffer, 0, 1, &scissor);
-    }
-
-    pipeline->bind(commandBuffer);
-
-    auto &reg = ECSRegistry::getRegistry();
-    const auto entityList = reg.getEntities();
-    for (const Entity e : entityList) {
-        auto &transform = reg.getComponent<impgine::Transform>(e);
-        auto &meshRenderer = reg.getComponent<impgine::MeshRenderer3D>(e);
-
-        // Get the mesh GPU resources
-        auto meshIt = meshCache.find(meshRenderer.mesh->modelPath);
-        if (meshIt == meshCache.end()) {
-            continue; // Skip if mesh not loaded
-        }
-        auto& meshResources = meshIt->second;
-
-        // Bind descriptor set for this mesh (includes its texture)
-        vkCmdBindDescriptorSets(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS, pipelineLayout, 0, 1, &meshResources.descriptorSets[imageIndex], 0, nullptr);
-
-        // Bind vertex and index buffers for this entity's mesh
-        VkBuffer vertexBuffers[] = {meshResources.vertexBuffer};
-        VkDeviceSize offsets[] = {0};
-        vkCmdBindVertexBuffers(commandBuffer, 0, 1, vertexBuffers, offsets);
-        vkCmdBindIndexBuffer(commandBuffer, meshResources.indexBuffer, 0, VK_INDEX_TYPE_UINT32);
-
-        // Build model matrix for this entity
-        glm::mat4 model = glm::mat4(1.0f);
-        model = glm::translate(model, transform.position);
-        model = glm::rotate(model, transform.rotation.x, glm::vec3(1.0f, 0.0f, 0.0f));
-        model = glm::rotate(model, transform.rotation.y, glm::vec3(0.0f, 1.0f, 0.0f));
-        model = glm::rotate(model, transform.rotation.z, glm::vec3(0.0f, 0.0f, 1.0f));
-        model = glm::scale(model, transform.scale);
-
-        // Push the model matrix via push constants
-        PushConstantData pushData{};
-        pushData.model = model;
-        vkCmdPushConstants(commandBuffer, pipelineLayout, VK_SHADER_STAGE_VERTEX_BIT, 0, sizeof(PushConstantData), &pushData);
-
-        vkCmdDrawIndexed(commandBuffer, static_cast<uint32_t>(meshResources.indices.size()), 1, 0, 0, 0);
-    }
-
-    // Draw UI on top
-    if (uiRenderer) {
-        uiRenderer->record(commandBuffer, *window, uiWindows, imageIndex);
-    }
-
-    vkCmdEndRenderPass(commandBuffer);
-
-    if (vkEndCommandBuffer(commandBuffer) != VK_SUCCESS) {
-        throw std::runtime_error("failed to record command buffer!");
-    }
-}
-
 void Engine::drawFrame() {
-    // Wait for the previous frame to finish
+    // Wait for the previous frame
     VkFence inFlightFence = swapChain->getInFlightFence(currentFrame);
     vkWaitForFences(device, 1, &inFlightFence, VK_TRUE, UINT64_MAX);
 
     uint32_t imageIndex;
-    // Use frame-based semaphore for acquire
     VkSemaphore imageAvailableSemaphore = swapChain->getImageAvailableSemaphore(currentFrame);
     auto result = swapChain->acquireNextImage(&imageIndex, currentFrame, imageAvailableSemaphore);
 
@@ -1660,14 +342,31 @@ void Engine::drawFrame() {
         throw std::runtime_error("failed to acquire swap chain image!");
     }
 
-    // Only reset the fence if we are submitting work
     vkResetFences(device, 1, &inFlightFence);
 
-    updateUniformBuffer(imageIndex);
+    // Update camera projection and view matrices
+    camera.setPerspectiveProjection(
+        glm::radians(45.0f),
+        swapChain->getSwapChainExtent().width / (float)swapChain->getSwapChainExtent().height,
+        0.1f,
+        100.0f
+    );
+    camera.updateViewMatrix();
 
-    // Reset and record command buffer
+    // Update uniform buffer - use imageIndex for correct buffer
+    UniformBufferObject ubo{};
+    ubo.view = camera.getView();
+    ubo.proj = camera.getProjection();
+    impgine::updateUniformBuffer(device, uniformBuffersMemory[imageIndex], &ubo, sizeof(ubo));
+
+    // Record command buffer
     vkResetCommandBuffer(commandBuffers[currentFrame], 0);
-    recordCommandBuffer(commandBuffers[currentFrame], imageIndex);
+    auto& registry = ECSRegistry::getRegistry();
+    impgine::recordCommandBuffer(commandBuffers[currentFrame], imageIndex, renderPass,
+                                 swapChainFramebuffers, swapChain->getSwapChainExtent(),
+                                 pipeline.get(), pipelineLayout,
+                                 &resourceManager->getMeshCache(), &registry,
+                                 uiRenderer.get(), window.get(), &uiWindows);
 
     // Submit command buffer
     VkSubmitInfo submitInfo{};
@@ -1703,62 +402,32 @@ void Engine::drawFrame() {
 }
 
 void Engine::recreateSwapChain() {
-    int width = 0, height = 0;
-    window->getFramebufferSize(&width, &height);
-    while (width == 0 || height == 0) {
-        window->getFramebufferSize(&width, &height);
-        window->waitEvents();
-    }
-
-    vkDeviceWaitIdle(device);
-
-    // Clean up old uniform buffers and descriptor pool
-    for (size_t i = 0; i < uniformBuffers.size(); i++) {
-        vkDestroyBuffer(device, uniformBuffers[i], nullptr);
-        vkFreeMemory(device, uniformBuffersMemory[i], nullptr);
-    }
-    
-    vkDestroyDescriptorPool(device, descriptorPool, nullptr);
-
-    cleanupSwapChain();
-
-    swapChain = std::make_unique<SwapChain>(device, physicalDevice, surface, *window);
-
-    createColorResources();
-    createDepthResources();
-    createRenderPass();
-    createFramebuffers();
-
-    // Recreate uniform buffers and descriptor resources for new swap chain
-    createUniformBuffers();
-    createDescriptorPool();
-    createDescriptorSets();
-
-    // Recreate pipeline since it depends on render pass
-    pipeline.reset();
-
-    PipelineConfigInfo pipelineConfig{};
-    Pipeline::defaultPipelineConfigInfo(pipelineConfig);
-    pipelineConfig.renderPass = renderPass;
-    pipelineConfig.pipelineLayout = pipelineLayout;
-    pipelineConfig.multisampleInfo.rasterizationSamples = msaaSamples;
-
     std::string vertShader = project.getFullPath(project.shadersPath + "/vert.spv");
     std::string fragShader = project.getFullPath(project.shadersPath + "/frag.spv");
-    pipeline = std::make_unique<Pipeline>(device, vertShader, fragShader, pipelineConfig);
+    std::string uiVert = project.getFullPath(project.shadersPath + "/ui.vert.spv");
+    std::string uiFrag = project.getFullPath(project.shadersPath + "/ui.frag.spv");
+    std::string fontPath = project.getFullPath("engine/ttf/arial.ttf");
 
-    // Recreate UI renderer with new render pass and extent
-    if (uiRenderer) {
-        uiRenderer.reset();
-        std::string uiVert = project.getFullPath(project.shadersPath + "/ui.vert.spv");
-        std::string uiFrag = project.getFullPath(project.shadersPath + "/ui.frag.spv");
-        uiRenderer = std::make_unique<UIRenderer>(device, physicalDevice, renderPass, pipelineLayout, commandPool, graphicsQueue, *swapChain, msaaSamples, uiVert, uiFrag);
+    // Release ownership from unique_ptrs so impgine::recreateSwapChain can delete them
+    SwapChain* swapChainPtr = swapChain.release();
+    Pipeline* pipelinePtr = pipeline.release();
+    UIRenderer* uiRendererPtr = uiRenderer.release();
 
-        // Reload font and create descriptor sets
-        std::string fontPath = project.getFullPath("engine/ttf/arial.ttf");
-        uiRenderer->getTextRenderer()->loadFont(fontPath, 48);
-        uiRenderer->createDescriptorSets();
-    }
+    impgine::recreateSwapChain(device, physicalDevice, surface, *window, swapChainPtr,
+                               colorImage, colorImageMemory, colorImageView,
+                               depthImage, depthImageMemory, depthImageView,
+                               renderPass, swapChainFramebuffers,
+                               uniformBuffers, uniformBuffersMemory,
+                               descriptorPool, descriptorSetLayout,
+                               &resourceManager->getMeshCache(), msaaSamples,
+                               &pipelinePtr, pipelineLayout, vertShader, fragShader,
+                               &uiRendererPtr, commandPool, graphicsQueue,
+                               uiVert, uiFrag, fontPath);
+
+    // Take ownership of new pointers
+    swapChain.reset(swapChainPtr);
+    pipeline.reset(pipelinePtr);
+    uiRenderer.reset(uiRendererPtr);
 }
 
 void Engine::framebufferResizeCallback(GLFWwindow* window, int width, int height) {
@@ -1769,172 +438,9 @@ void Engine::framebufferResizeCallback(GLFWwindow* window, int width, int height
 }
 
 void Engine::mouseButtonCallback(GLFWwindow* window, int button, int action, int mods) {
-    (void)mods;
     auto app = reinterpret_cast<Engine*>(glfwGetWindowUserPointer(window));
-
-    if (button == GLFW_MOUSE_BUTTON_LEFT && action == GLFW_PRESS) {
-        double xpos, ypos;
-        glfwGetCursorPos(window, &xpos, &ypos);
-
-        // Convert to framebuffer coordinates
-        int fbWidth, fbHeight;
-        glfwGetFramebufferSize(window, &fbWidth, &fbHeight);
-        int winWidth, winHeight;
-        glfwGetWindowSize(window, &winWidth, &winHeight);
-
-        xpos = xpos * fbWidth / winWidth;
-        ypos = ypos * fbHeight / winHeight;
-
-        if (app->uiRenderer) {
-            app->uiRenderer->handleMouseClick(xpos, ypos, app->uiWindows);
-        }
-    }
-}
-
-VKAPI_ATTR VkBool32 VKAPI_CALL
-Engine::debugCallback(VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity,
-                      VkDebugUtilsMessageTypeFlagsEXT messageType,
-                      const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData, void* pUserData) {
-    (void)messageSeverity;
-    (void)messageType;
-    (void)pUserData;
-
-    std::cerr << "validation layer: " << pCallbackData->pMessage << std::endl;
-
-    return VK_FALSE;
-}
-
-void Engine::processInput(float deltaTime) {
-    const float moveSpeed = 5.0f; // units per second
-    
-    // WASD movement (using ZQSD for AZERTY keyboards, but we'll use WASD)
-    if (window->isKeyPressed(GLFW_KEY_W)) {
-        camera.moveForward(moveSpeed * deltaTime);
-    }
-    if (window->isKeyPressed(GLFW_KEY_S)) {
-        camera.moveBackward(moveSpeed * deltaTime);
-    }
-    if (window->isKeyPressed(GLFW_KEY_A)) {
-        camera.moveLeft(moveSpeed * deltaTime);
-    }
-    if (window->isKeyPressed(GLFW_KEY_D)) {
-        camera.moveRight(moveSpeed * deltaTime);
-    }
-    
-    // Space and Shift for up/down movement
-    if (window->isKeyPressed(GLFW_KEY_SPACE)) {
-        camera.moveUp(moveSpeed * deltaTime);
-    }
-    if (window->isKeyPressed(GLFW_KEY_LEFT_SHIFT)) {
-        camera.moveDown(moveSpeed * deltaTime);
-    }
-    
-    // ESC to release mouse capture
-    if (window->isKeyPressed(GLFW_KEY_ESCAPE)) {
-        if (mouseCaptured) {
-            mouseCaptured = false;
-            window->setCursorInputMode(GLFW_CURSOR_NORMAL);
-        }
-    }
-}
-
-void Engine::handleMouseMovement() {
-    if (!mouseCaptured) return;
-    
-    double xpos, ypos;
-    window->getCursorPos(&xpos, &ypos);
-    
-    if (firstMouse) {
-        lastMouseX = xpos;
-        lastMouseY = ypos;
-        firstMouse = false;
-    }
-    
-    double xoffset = xpos - lastMouseX;
-    double yoffset = ypos - lastMouseY; // Normal direction for intuitive mouse look
-    
-    lastMouseX = xpos;
-    lastMouseY = ypos;
-    
-    const float sensitivity = 0.002f; // Adjust as needed
-    camera.rotateYaw(static_cast<float>(xoffset * sensitivity));
-    camera.rotatePitch(static_cast<float>(yoffset * sensitivity));
-}
-
-void Engine::handleUIInput() {
-    // Get mouse position in screen coordinates
-    double mx, my;
-    window->getCursorPos(&mx, &my);
-
-    // Scale mouse coordinates to framebuffer coordinates for UI hit testing
-    int windowWidth = 0, windowHeight = 0;
-    int fbWidth = 0, fbHeight = 0;
-    window->getWindowSize(&windowWidth, &windowHeight);
-    window->getFramebufferSize(&fbWidth, &fbHeight);
-
-    float scaleX = (windowWidth > 0) ? static_cast<float>(fbWidth) / static_cast<float>(windowWidth) : 1.0f;
-    float scaleY = (windowHeight > 0) ? static_cast<float>(fbHeight) / static_cast<float>(windowHeight) : 1.0f;
-
-    float mouseX = static_cast<float>(mx) * scaleX;
-    float mouseY = static_cast<float>(my) * scaleY;
-
-    // Check if mouse is over any UI panel
-    bool mouseOverUI = uiRenderer->getLayout().isPointOverUI(mouseX, mouseY, uiWindows);
-
-    // Handle mouse capture for 3D viewport navigation
-    if (!mouseCaptured) {
-        if (window->isMouseButtonPressed(GLFW_MOUSE_BUTTON_LEFT)) {
-            if (!mouseOverUI) {
-                mouseCaptured = true;
-                window->setCursorInputMode(GLFW_CURSOR_DISABLED);
-
-                // Center cursor
-                int fbw = 0, fbh = 0;
-                window->getFramebufferSize(&fbw, &fbh);
-                window->setCursorPos(fbw / 2.0, fbh / 2.0);
-                firstMouse = true;
-            }
-        }
-    }
-
-    // Handle UI window dragging (optional feature)
-    // Currently disabled, but can be re-enabled for floating windows
-}
-
-bool Engine::compileShader(const std::string& sourcePath, const std::string& outputPath, const std::string& shaderType) {
-    std::string command = "glslc " + sourcePath + " -o " + outputPath;
-    std::cout << "Compiling " << shaderType << " shader: " << sourcePath << std::endl;
-    int result = std::system(command.c_str());
-    if (result != 0) {
-        std::cerr << "Failed to compile shader: " << sourcePath << std::endl;
-        return false;
-    }
-    std::cout << "Successfully compiled " << shaderType << " shader to " << outputPath << std::endl;
-    return true;
-}
-
-void Engine::compileShaders() {
-    std::cout << "Compiling shaders..." << std::endl;
-    std::string shaderPath = project.getFullPath(project.shadersPath);
-    std::string vertSource = shaderPath + "/shader.vert";
-    std::string fragSource = shaderPath + "/shader.frag";
-    std::string vertOutput = shaderPath + "/vert.spv";
-    std::string fragOutput = shaderPath + "/frag.spv";
-    // UI shaders
-    std::string uiVertSrc = shaderPath + "/ui.vert";
-    std::string uiFragSrc = shaderPath + "/ui.frag";
-    std::string uiVertOut = shaderPath + "/ui.vert.spv";
-    std::string uiFragOut = shaderPath + "/ui.frag.spv";
-
-    bool vertSuccess = compileShader(vertSource, vertOutput, "vertex");
-    bool fragSuccess = compileShader(fragSource, fragOutput, "fragment");
-    bool uiVertSuccess = compileShader(uiVertSrc, uiVertOut, "vertex");
-    bool uiFragSuccess = compileShader(uiFragSrc, uiFragOut, "fragment");
-
-    if (!vertSuccess || !fragSuccess || !uiVertSuccess || !uiFragSuccess) {
-        throw std::runtime_error("Failed to compile shaders");
-    }
-    std::cout << "All shaders compiled successfully!" << std::endl;
+    InputManager::mouseButtonCallback(window, button, action, mods,
+                                     app->uiRenderer.get(), app->uiWindows);
 }
 
 }  // namespace impgine
